requirements.txt
django==5.1.2
djangorestframework==3.15.2
python-dotenv==1.0.1
redis==5.2.0
celery==5.4.0
gunicorn==23.0.0
psycopg2-binary==2.9.9
